pyproject.toml
[project]
name = "t3clone-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.115.12",
    "psycopg2>=2.9.10",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.24",
    "websockets>=15.0.1",
]
